docs : delete comments

+++ src/Client.cpp
@@ -48,7 +48,6 @@ std::vector<Message> Client::readData()
 
     char buffer[IN_BUFFER_SIZE];
 
-    //memset(buffer, 0, sizeof(buffer));
     bzero(buffer, IN_BUFFER_SIZE);
 
     int received = recv(this->socket, buffer, sizeof(buffer), 0);
@@ -133,7 +132,7 @@ void Client::writeOff()
 
 void Client::sendData()
 {
-    size_t sentLength;
+    ssize_t sentLength;
 
     sentLength = send(socket, outBuffer.c_str(), outBuffer.size(), MSG_DONTWAIT);
 
@@ -141,7 +140,7 @@ void Client::sendData()
     {
         throw std::runtime_error("send error");
     }
-    else if (sentLength != outBuffer.size()) // 다 못 보냈을 때
+    else if (sentLength != (ssize_t)outBuffer.size()) // 다 못 보냈을 때
     {
         this->outBuffer.erase(0, sentLength);
     }
